add linter config
