+++ .stickler.yml
@@ -1,0 +1,3 @@
+linters:
+  rubocop:
+    display_cop_names: true
